refactor: delete 2fa auth
