+++ prisma/schema.prisma
@@ -13,22 +13,14 @@ datasource db {
     url      = env("DATABASE_URL")
 }
 
-enum UserRole {
-    ADMIN
-    USER
-}
-
 model User {
-    id                    String                 @id @default(cuid())
-    name                  String?
-    email                 String                 @unique
-    emailVerified         DateTime?
-    image                 String?
-    password              String?
-    role                  UserRole               @default(USER)
-    accounts              Account[]
-    isTwoFactorEnabled    Boolean                @default(true)
-    twoFactorConfirmation TwoFactorConfirmation?
+    id            String    @id @default(cuid())
+    name          String?
+    email         String    @unique
+    emailVerified DateTime?
+    image         String?
+    password      String?
+    accounts      Account[]
 
     createdAt DateTime @default(now())
     updatedAt DateTime @updatedAt
@@ -55,15 +47,6 @@ model Account {
     @@id([provider, providerAccountId])
 }
 
-model VerificationToken {
-    id      String   @id @default(cuid())
-    email   String
-    token   String   @unique
-    expires DateTime
-
-    @@unique([email, token])
-}
-
 model PasswordResetToken {
     id      String   @id @default(cuid())
     email   String
@@ -71,21 +54,4 @@ model PasswordResetToken {
     expires DateTime
 
     @@unique([email, token])
-}
-
-model TwoFactorToken {
-    id      String   @id @default(cuid())
-    email   String
-    token   String   @unique
-    expires DateTime
-
-    @@unique([email, token])
-}
-
-model TwoFactorConfirmation {
-    id     String @id @default(cuid())
-    userId String
-    user   User   @relation(fields: [userId], references: [id], onDelete: Cascade)
-
-    @@unique([userId])
 }
