+++ src/app/(unauthorized)/(sign)/reset-password/components/ResetForm.tsx
@@ -3,11 +3,12 @@
 import { resetPassword } from '@/app/api/auth/resetPassword/actions';
 import { ResetPasswordDto } from '@/app/api/auth/resetPassword/dto';
 import { Button } from '@/components/ui/button';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { useErrorHandler } from '@/hooks/useErrorHandler';
 import { useResetPasswordForm } from '@/hooks/useResetPasswordForm';
+import Link from 'next/link';
 import React, { useTransition } from 'react';
 
 interface Props {
@@ -31,41 +32,50 @@ export const ResetForm: React.FC<Props> = ({ token }) => {
 	return (
 		<Card>
 			<CardHeader>
-				<CardTitle>Reset Password</CardTitle>
+				<CardTitle className='text-center'>
+					<h1>Forgot password?</h1>
+				</CardTitle>
+				<CardDescription className='text-center'>
+					Enter the email address associated with account.
+				</CardDescription>
 			</CardHeader>
 			<CardContent>
 				<Form {...form}>
 					<form onSubmit={form.handleSubmit(handleSubmit)} className='min-w-[22.5rem] space-y-4'>
-						<div>
-							<FormField
-								control={form.control}
-								name='password'
-								render={({ field }) => (
-									<FormItem>
-										<FormLabel>New password</FormLabel>
-										<FormControl>
-											<Input type='password' {...field} />
-										</FormControl>
-										<FormMessage />
-									</FormItem>
-								)}
-							/>
-							<FormField
-								control={form.control}
-								name='confirmPassword'
-								render={({ field }) => (
-									<FormItem>
-										<FormLabel>Confirm password</FormLabel>
-										<FormControl>
-											<Input type='password' {...field} />
-										</FormControl>
-										<FormMessage />
-									</FormItem>
-								)}
-							/>
-							<div className='flex items-center justify-center pt-6'>
-								<Button disabled={isPending}>Reset</Button>
-							</div>
+						<FormField
+							control={form.control}
+							name='password'
+							render={({ field }) => (
+								<FormItem>
+									<FormControl>
+										<Input type='password' placeholder='Enter new password' {...field} />
+									</FormControl>
+									<FormMessage />
+								</FormItem>
+							)}
+						/>
+						<FormField
+							control={form.control}
+							name='confirmPassword'
+							render={({ field }) => (
+								<FormItem>
+									<FormControl>
+										<Input type='password' placeholder='Confirm password' {...field} />
+									</FormControl>
+									<FormMessage />
+								</FormItem>
+							)}
+						/>
+						<p className='pt-6 text-center text-sm'>
+							Back to{' '}
+							<Link href='/sign-in' className='text-link'>
+								Sign in
+							</Link>
+						</p>
+						<div className='flex items-center justify-center'>
+							<Button disabled={isPending} size='lg' className='w-full'>
+								Reset password
+							</Button>
 						</div>
 					</form>
 				</Form>
