+++ src/app/api/auth/service.ts
@@ -8,13 +8,8 @@ import { SignInDto } from './sign-in/dto';
 import { SignUpDto } from './sign-up/dto';
 import bcrypt from 'bcrypt';
 import { getUserByEmail } from '../user/service';
-import {
-	generateTwoFactorToken,
-	getTwoFactorConfirmationByUserId,
-	getTwoFactorTokenByToken,
-} from './twoFactor/actions';
 import { generatePasswordResetToken, getPasswordResetTokenByToken } from './forgot-password/actions';
-import { sendPasswordResetTokenEmail, sendTwoFactorTokenEmail } from '../mail/actions';
+import { sendPasswordResetTokenEmail } from '../mail/actions';
 
 export const signUp = async ({ name, email, password }: SignUpDto) => {
 	const existingUser = await getUserByEmail(email);
@@ -31,8 +26,6 @@ export const signUp = async ({ name, email, password }: SignUpDto) => {
 		},
 	});
 
-	await twoFactorToken(email);
-
 	return NextResponse.json(null, { status: 201 });
 };
 
@@ -42,8 +35,8 @@ export const signIn = async ({ email, password }: SignInDto) => {
 		throw new RequestError('User not found', 400);
 	}
 
-	if (user.accounts) {
-		return NextResponse.json(null, { status: 200 });
+	if (user.accounts.length) {
+		throw new RequestError('Invalid provider', 500);
 	}
 
 	if (user.password) {
@@ -53,29 +46,6 @@ export const signIn = async ({ email, password }: SignInDto) => {
 		}
 	}
 
-	const existingTwoFactorConfirmation = await getTwoFactorConfirmationByUserId(user.id);
-	if (!existingTwoFactorConfirmation) {
-		await twoFactorToken(email);
-	}
-
-	return NextResponse.json(null, { status: 200 });
-};
-
-const twoFactorToken = async (email: string) => {
-	const { token } = await generateTwoFactorToken(email);
-	await sendTwoFactorTokenEmail(email, token);
-};
-
-export const twoFactorCheckCode = async (token: string) => {
-	const existingToken = await getTwoFactorTokenByToken(token);
-	if (!existingToken) {
-		throw new RequestError('Invalid token', 500);
-	}
-
-	if (existingToken.expires < new Date()) {
-		throw new RequestError('Token has expired', 500);
-	}
-
 	return NextResponse.json(null, { status: 200 });
 };
 
@@ -83,6 +53,10 @@ export const forgotPassword = async (email: string) => {
 	const user = await getUserByEmail(email);
 	if (!user) {
 		throw new RequestError('User not exists', 500);
+	}
+
+	if (user.accounts.length) {
+		throw new RequestError('Invalid provider', 500);
 	}
 
 	const token = await generatePasswordResetToken(email);
