+++ src/app/api/auth/sign-in/actions.ts
@@ -6,9 +6,10 @@ import { providerMap } from '@/auth.config';
 import { signIn } from '../service';
 import { env } from '@/configs/env';
 
-export const signInWithCredentials = async (dto: SignInDto) => {
+export const signInWithCredentials = async (dto: SignInDto, callbackUrl?: string) => {
 	try {
 		await signIn(dto);
+		await _signIn('credentials', { ...dto, redirectTo: callbackUrl || env.AUTH_DEFAULT_REDIRECT_URL });
 	} catch (error) {
 		throw error;
 	}
