--- src/app/api/auth/twoFactor/dto.ts
@@ -1,6 +1,0 @@
-import { z } from 'zod';
-
-export type TwoFactorCodeDto = z.infer<typeof twoFactorCodeDto>;
-export const twoFactorCodeDto = z.object({
-	token: z.string(),
-});
