+++ src/app/api/mail/actions.ts
@@ -10,12 +10,3 @@ export const sendPasswordResetTokenEmail = async (email: string, token: string) 
 		html: `<p>To reset password follow this link: ${resetLink}</p>`,
 	});
 };
-
-export const sendTwoFactorTokenEmail = async (email: string, token: string) => {
-	await resend.emails.send({
-		from: 'onboarding@resend.dev',
-		to: email,
-		subject: '2FA Code',
-		html: `<p>Your 2FA code: ${token}</p>`,
-	});
-};
